fix/dashboard-tanstack-query
Dashboard tanstack query fixing
